--- packages/gw-org2/src/init-org.ts
@@ -1,91 +1,0 @@
-require('./env');
-import util from 'util';
-import { getLogger } from '@fabric-es/gateway-lib';
-import { enrollAdmin, prepareOrgKeys } from '@fabric-es/operator';
-import { Wallet, Wallets } from 'fabric-network';
-import rimraf from 'rimraf';
-
-const logger = getLogger('[gw-org1] init.js');
-
-void (async () => {
-  rimraf(`${process.env.WALLET}/${process.env.ORG_ADMIN_ID}.id`, async error => {
-    if (error) {
-      logger.error(error);
-      process.exit(1);
-    }
-    logger.info(`${process.env.WALLET}/${process.env.ORG_ADMIN_ID}.id is removed`);
-
-    let wallet: Wallet;
-    try {
-      wallet = await Wallets.newFileSystemWallet(process.env.WALLET);
-    } catch (e) {
-      logger.error(e);
-      process.exit(1);
-    }
-
-    await enrollAdmin({
-      enrollmentID: process.env.ORG_ADMIN_ID,
-      enrollmentSecret: process.env.ORG_ADMIN_SECRET,
-      mspId: process.env.MSPID,
-      caName: process.env.CA_NAME,
-      connectionProfile: process.env.CONNECTION_PROFILE,
-      wallet
-    }).then(result => {
-      console.log(result);
-      logger.info(util.format('enrollAdmin successfully, %j', result));
-    }).catch(error => {
-      console.error(error);
-      logger.error(util.format('fail to enrollAdmin, %j', error));
-      process.exit(1);
-    });
-  });
-
-  await new Promise(resolve => setTimeout(resolve, 1000));
-
-  rimraf(`${process.env.WALLET}/${process.env.CA_ENROLLMENT_ID_ADMIN}.id`, async error => {
-    if (error) {
-      logger.error(error);
-      process.exit(1);
-    }
-    logger.info(`${process.env.WALLET}/${process.env.CA_ENROLLMENT_ID_ADMIN}.id is removed`);
-
-    let wallet: Wallet;
-    try {
-      wallet = await Wallets.newFileSystemWallet(process.env.WALLET);
-    } catch (e) {
-      logger.error(e);
-      process.exit(1);
-    }
-
-    await enrollAdmin({
-      enrollmentID: process.env.CA_ENROLLMENT_ID_ADMIN,
-      enrollmentSecret: process.env.CA_ENROLLMENT_SECRET_ADMIN,
-      mspId: process.env.MSPID,
-      caName: process.env.CA_NAME,
-      connectionProfile: process.env.CONNECTION_PROFILE,
-      wallet
-    }).then(result => {
-      console.log(result);
-      logger.info(util.format('enrollCaAdmin successfully, %j', result));
-    }).catch(error => {
-      console.error(error);
-      logger.error(util.format('fail to enrollCaAdmin, %j', error));
-      process.exit(1);
-    });
-  });
-
-  await new Promise(resolve => setTimeout(resolve, 1000));
-
-  await prepareOrgKeys({
-    keyPath: process.env.ORGKEY,
-    curve: process.env.ORGKEY_CURVE || undefined,
-  }).then(result => {
-    console.log(result);
-    logger.info(util.format('prepareOrgKeys successfully, %j', result));
-    process.exit(0);
-  }).catch(error => {
-    console.error(error);
-    logger.error(util.format('fail to prepareOrgKeys, %j', error));
-    process.exit(1);
-  });
-})();
