+++ packages/gw-org3/package.json
@@ -19,7 +19,7 @@
     "start-gateway": "node ./dist/app.js",
     "enrollAdmin": "node ./dist/enrollAdmin.js",
     "enrollCaAdmin": "node ./dist/enrollCaAdmin.js",
-    "init-org": "node ./dist/init-org.js",
+    "prepareKeyPair": "node ./dist/prepareKeyPair.js",
     "gen-resolvers-type": "graphql-codegen --config codegen.yml",
     "lint": "eslint --max-warnings 10 'src/**/*.{js,ts}'",
     "test:multi": "yarn build && ./node_modules/.bin/jest multi.test",
